Refactor code to facilitate reuse and use in serializable streams.
