+++ src/test/java/io/cdap/plugin/gcp/publisher/source/PubSubReceiverMessagePullingTest.java
@@ -73,7 +73,7 @@ public class PubSubReceiverMessagePullingTest {
   @Mock
   UnaryCallable<AcknowledgeRequest, Empty> acknowledgeCallable;
   @Mock
-  PubSubReceiver.BackoffConfig backoffConfig;
+  BackoffConfig backoffConfig;
   @Mock
   ScheduledThreadPoolExecutor executor;
